updated price to int

+++ SmartArt-Backend/src/test/java/ca/mcgill/ecse321/smartart/dao/TestSmartArtPersistence.java
@@ -216,7 +216,7 @@ public class TestSmartArtPersistence {
 	public void testPersistAndLoadPosting() {
 		int id = 7464;
 		String title = "Nu couche";
-		double price = 100000000;
+		int price = 100000000;
 		String email = "amy@mail.com";
 		String name = "amy";
 		String galName = "Guginhiem";
